Re-work how struct ownership is handled, simplify view binding
Now, we bind all non-POD Swift structs as managed reference type
subclasses of SwiftStruct. This allows us to ensure ownership and
lifetimes are properly handled by not passing ownership, but
keeping it until disposed or finalized. This simplifies the API
and makes things work like regular C# objects, at the potential
cost of some performance. For now, let's keep it simple, and we
can revisit performance later if necessary.