Rework SwiftType system to handle nullable reference types
- Breaks backward dep from Swift.Interop -> SwiftUI namespace
- Finishes the transparent optional bridging
- Adds new tests, all passing

A byproduct of this work is that types are bridged to Swift with
a custom attribute to get the SwiftType instead of having to declare
a public static SwiftType property. In the general case, bindings
can simply add a [SwiftImport] attribute with the library path
from which they are to be imported. That should be enough to handle
generic and non-generic types. For opaque return types, or types with
a special mangling, you must specify the mangled name (for opaque return
types, this is the mangled name for the descriptor).